Update to handle EODatasets thumbnail writing and geometry inheritance
