+++ datacube_alchemist/_utils.py
@@ -1,10 +1,16 @@
-from eodatasets3.model import DatasetDoc, ProductDoc
-from eodatasets3.properties import StacPropertyView
-from datacube.model import Dataset
-
+import json
+from pathlib import Path
 from typing import Dict
 
-from datacube.virtual import Transformation, Measurement
+from datacube.model import Dataset
+from datacube.virtual import Measurement, Transformation
+from eodatasets3 import DatasetAssembler, serialise
+from eodatasets3.model import DatasetDoc, ProductDoc
+from eodatasets3.properties import StacPropertyView
+from eodatasets3.scripts.tostac import dc_to_stac, json_fallback
+from eodatasets3.verify import PackageChecksum
+
+from datacube_alchemist.settings import AlchemistTask
 
 
 class FakeTransformation(Transformation):
@@ -17,6 +23,46 @@ class FakeTransformation(Transformation):
 
     def compute(self, data) -> Dataset:
         return data
+
+
+def _write_thumbnail(task: AlchemistTask, dataset_assembler: DatasetAssembler):
+    if task.settings.output.preview_image is not None:
+        dataset_assembler.write_thumbnail(*task.settings.output.preview_image)
+    elif task.settings.output.preview_image_singleband is not None:
+        dataset_assembler.write_thumbnail_singleband(
+            **task.settings.output.preview_image_singleband
+        )
+
+
+def _write_stac(
+    metadata_path: Path,
+    task: AlchemistTask,
+    dataset_assembler: DatasetAssembler,
+):
+    out_dataset = serialise.from_path(metadata_path)
+    stac_path = Path(str(metadata_path).replace("odc-metadata.yaml", "stac-item.json"))
+    stac = dc_to_stac(
+        out_dataset,
+        metadata_path,
+        stac_path,
+        stac_path.root,
+        task.settings.output.explorer_url,
+        False,
+    )
+    with stac_path.open("w") as f:
+        json.dump(stac, f, default=json_fallback)
+    dataset_assembler.add_accessory_file("metadata:stac", stac_path)
+
+    # dataset_assembler._checksum.write(dataset_assembler._accessories["checksum:sha1"])
+    # Need a new checksummer because EODatasets is insane
+    checksummer = PackageChecksum()
+    checksum_file = (
+        dataset_assembler._dataset_location
+        / dataset_assembler._accessories["checksum:sha1"].name
+    )
+    checksummer.read(checksum_file)
+    checksummer.add_file(stac_path)
+    checksummer.write(checksum_file)
 
 
 def _munge_dataset_to_eo3(ds: Dataset) -> DatasetDoc:
@@ -34,7 +80,11 @@ def _munge_dataset_to_eo3(ds: Dataset) -> DatasetDoc:
     # Wrap properties to avoid typos and the like
     properties = StacPropertyView(ds.metadata_doc.get("properties", {}))
     return DatasetDoc(
-        id=ds.id, product=product, crs=ds.crs.crs_str, properties=properties
+        id=ds.id,
+        product=product,
+        crs=str(ds.crs),
+        properties=properties,
+        geometry=ds.extent,
     )
 
 
@@ -53,9 +103,7 @@ def _convert_eo_plus(ds) -> DatasetDoc:
         }
     )
     product = ProductDoc(name=ds.type.name)
-    return DatasetDoc(
-        id=ds.id, product=product, crs=ds.crs.crs_str, properties=properties
-    )
+    return DatasetDoc(id=ds.id, product=product, crs=str(ds.crs), properties=properties)
 
 
 def _convert_eo(ds) -> DatasetDoc:
@@ -70,6 +118,4 @@ def _convert_eo(ds) -> DatasetDoc:
         }
     )
     product = ProductDoc(name=ds.type.name)
-    return DatasetDoc(
-        id=ds.id, product=product, crs=ds.crs.crs_str, properties=properties
-    )
+    return DatasetDoc(id=ds.id, product=product, crs=str(ds.crs), properties=properties)
